Add files via upload
Assignment Details: 
Create a basic webpage that uses JavaScript to calculate the volume of 5 identical boxes.  Have variables to hold the length, width and height of a box.  Besides the calculations, the JavaScript code should also write to the webpage the volume of a single box as well as the total volume of all 5 boxes combined.  Make sure to give the user the info in complete sentences so they understand what the numbers represent.

For this lab and all future labs and projects take print screens of both your code and the code running in the browser window and paste them into a Word Document.  Turn the Word Document into Canvas.